* 进行安全处理

+++ langx-java/src/main/java/com/jn/langx/util/Numbers.java
@@ -286,10 +286,10 @@ public class Numbers {
             return number.longValue() == 0L;
         }
         if (number instanceof Float) {
-            return ((Float)number).floatValue() == 0F;
+            return (Float) number == 0F;
         }
         if (number instanceof Double) {
-            return ((Double)number).doubleValue() == 0D;
+            return (Double) number == 0D;
         }
         if (number instanceof AtomicInteger) {
             return ((AtomicInteger) number).get() == 0;
